finish video >> redis query

+++ util/uuid.go
@@ -4,6 +4,6 @@ import (
 	"github.com/godruoyi/go-snowflake"
 )
 
-func UniqueID() uint64 {
-	return snowflake.ID()
+func UniqueID() int64 {
+	return int64(snowflake.ID())
 }
